add controllers for all routes
