+++ index.js
@@ -1,6 +1,5 @@
 const express = require("express");
 const cors = require("cors");
-const bodyParser = require("body-parser");
 require("dotenv").config();
 const app = express();
 
@@ -14,10 +13,10 @@ const notificationRouter = require("./routes/notifications.route");
 const profileRoute = require("./routes/profile.route");
 
 app.use(cors());
-app.use(bodyParser.json());
-const urlencodedParser = bodyParser.urlencoded({ extended: false });
+app.use(express.json());
 
 const initializedDBConnection = require("./db/db.connect");
+const routeHandler = require("./middlewares/routeHandler");
 initializedDBConnection();
 
 app.use("/auth", authRouter);
@@ -25,19 +24,13 @@ app.use("/notifications", authVerify, notificationRouter);
 app.use("/user", authVerify, userRouter);
 app.use("/posts", authVerify, postRouter);
 app.use("/followunfollow", authVerify, followRouter);
-app.use("/profile", authVerify, profileRoute);
+app.use("/profile", profileRoute);
 
 app.get("/", (req, res) => {
-  res.send("Hello Social Media");
+  res.json({ success: true, message: "Hello Social Media" });
 });
 
-// route 404 do not move
-app.use((req, res) => {
-  res.status(404).json({
-    success: false,
-    message: "route not found on server",
-  });
-});
+app.use(routeHandler);
 
 app.listen(process.env.PORT || 7000, () => {
   console.log("Server Started");
