+++ middlewares/routeHandler.js
@@ -1,0 +1,10 @@
+// route 404 do not move
+
+const routeHandler = (req, res) => {
+  res.status(404).json({
+    success: false,
+    message: "route not found on server",
+  });
+};
+
+module.exports = routeHandler;
